requirements.txt
google-adk
ag-ui-adk

fastapi
uvicorn[standard]

python-dotenv
httpx
pydantic
sse-starlette